refactor(portal): extract Portal.view.content.Component timeline base (#12208)
