+++ apps/portal/view/news/tickets/Component.mjs
@@ -1,4 +1,4 @@
-import ContentComponent from '../../../../../src/app/content/Component.mjs';
+import ContentComponent from '../../content/Component.mjs';
 import {marked}         from '../../../../../node_modules/marked/lib/marked.esm.js';
 
 const
@@ -42,29 +42,11 @@ class Component extends ContentComponent {
          */
         commitsUrl: 'https://github.com/neomjs/neo/commit/',
         /**
-         * @member {Object} domListeners
-         */
-        domListeners: {
-            resize: 'onResize'
-        },
-        /**
          * @member {String} repoUserUrl='https://github.com/'
          */
-        repoUserUrl: 'https://github.com/',
-        /**
-         * @member {Boolean} updateSectionsStore=false
-         */
-        updateSectionsStore: false
+        repoUserUrl: 'https://github.com/'
     }
 
-    /**
-     * @member {Intl.DateTimeFormat|null} #dateTimeFormatHistory=null
-     */
-    #dateTimeFormatHistory = null
-    /**
-     * @member {Intl.DateTimeFormat|null} #dateTimeFormatToday=null
-     */
-    #dateTimeFormatToday = null
     /**
      * Temporary storage for the structured timeline data extracted during the parsing phase.
      * This array is populated by `renderTimeline` and `modifyMarkdown` and then assigned
@@ -73,50 +55,6 @@ class Component extends ContentComponent {
      * @private
      */
     timelineData = null
-
-    /**
-     * @param {Object} config
-     */
-    construct(config) {
-        super.construct(config);
-        this.getStateProvider().setData('contentComponentId', this.id)
-    }
-
-    /**
-     * @param {String} isoString
-     * @returns {String}
-     */
-    formatTimestamp(isoString) {
-        if (!isoString) return '';
-
-        let me      = this,
-            date    = new Date(isoString),
-            now     = new Date(),
-            isToday = date.toDateString() === now.toDateString();
-
-        if (isToday) {
-            if (!me.#dateTimeFormatToday) {
-                me.#dateTimeFormatToday = new Intl.DateTimeFormat('default', {
-                    hour  : 'numeric',
-                    minute: 'numeric'
-                })
-            }
-
-            return me.#dateTimeFormatToday.format(date)
-        }
-
-        if (!me.#dateTimeFormatHistory) {
-            me.#dateTimeFormatHistory = new Intl.DateTimeFormat('default', {
-                day   : 'numeric',
-                hour  : 'numeric',
-                minute: 'numeric',
-                month : 'short',
-                year  : 'numeric'
-            })
-        }
-
-        return me.#dateTimeFormatHistory.format(date)
-    }
 
     /**
      * @param {Object} data
@@ -177,43 +115,6 @@ class Component extends ContentComponent {
         badgesHtml += '</div>';
 
         return badgesHtml
-    }
-
-    /**
-     * @param {Object} record
-     * @returns {Promise<void>}
-     */
-    async doFetchContent(record) {
-        let me         = this,
-            {windowId} = me,
-            content, data, path;
-
-        path = me.getContentPath(record);
-
-        if (record.isLeaf && path) {
-            data    = await fetch(path);
-            content = await data.text();
-
-            me.value = content;
-
-            me.toggleCls('lab', record.name?.startsWith('Lab:'));
-
-            Neo.main.addon.IntersectionObserver.observe({
-                disconnect: true,
-                id        : me.id,
-                observe   : ['.neo-timeline-item[data-record-id]'],
-                windowId
-            })
-        }
-    }
-
-    /**
-     * @param {Object} record
-     * @param {String} record.path
-     * @returns {String|null}
-     */
-    getContentPath({path}) {
-        return path ? Neo.config.basePath + path : null
     }
 
     /**
@@ -402,13 +303,6 @@ ${fullHtml}
     }
 
     /**
-     * @param {Object} data
-     */
-    onResize(data) {
-        this.fire('toggleSummary')
-    }
-
-    /**
      * Parses the custom "Timeline" markdown section.
      *
      * Expects a line-based format generated by the build process:
@@ -439,7 +333,7 @@ ${fullHtml}
                 id = `timeline-${me.record.id}-${me.timelineData.length + 1}`;
 
                 me.timelineData.push({
-                    id   : id,
+                    id,
                     image: repoUserUrl + currentUser + '.png',
                     name : `Comment (${currentUser})`,
                     tag  : 'comment'
@@ -539,9 +433,9 @@ ${fullHtml}
                 }
 
                 me.timelineData.push({
-                    color: color, // Pass resolved hex color
-                    icon : icon,
-                    id   : id,
+                    color, // Pass resolved hex color
+                    icon,
+                    id,
                     name : entryName,
                     tag  : 'event'
                 });
