+++ resources/scss/src/apps/portal/content/CanvasWrapper.scss
